chore: refactor api.ts by extracting actual endpoint code to a new file for better ability to unit test
